Change JSSVCB key to int64 type

+++ modules/dns_proxy/dns_proxy_js_record_svcb.go
@@ -12,7 +12,7 @@ func NewJSSVCBKeyValue(kv dns.SVCBKeyValue) (map[string]interface{}, error) {
 	key := kv.Key()
 
 	jsKv := map[string]interface{}{
-		"Key": uint16(key),
+		"Key": int64(key),
 	}
 
 	switch v := kv.(type) {
